requirements.txt
Babel==2.4.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.6
django-phonenumber-field==1.2.0
django-toolbelt==0.0.1
gunicorn==19.6.0
html5lib==1.0b8
oauthlib==2.0.2
olefile==0.44
phonenumberslite==8.3.3
Pillow==4.0.0
psycopg2==2.7.1
PyJWT==1.4.2
PyPDF2==1.26.0
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0
static3==0.7.0
webencodings==0.5.1
wheel==0.26.0
whitenoise==3.3.0
xhtml2pdf==0.0.6
